+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Really fix opening Private window in Firefox. Thanks to @mpsijm for fix #134
 - macOS: Don't exit on lost focus if Settings or About dialog are still open
 - Windows: Fix crashing when Zoom is installed
+- Include also locked Firefox profiles. Might help with #81
 
 ## [0.5.1] - 2024-01-31