+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.3] - 2024-02-21
+
+### Fixed
+- Fix parsing Firefox containers for latest Firefox version
+
 ## [0.5.2] - 2024-02-08
 
 ### Fixed
@@ -242,7 +247,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial Release
 
-[unreleased]: https://github.com/Browsers-software/browsers/compare/0.5.2...HEAD
+[unreleased]: https://github.com/Browsers-software/browsers/compare/0.5.3...HEAD
+[0.5.3]: https://github.com/Browsers-software/browsers/releases/tag/0.5.3
 [0.5.2]: https://github.com/Browsers-software/browsers/releases/tag/0.5.2
 [0.5.1]: https://github.com/Browsers-software/browsers/releases/tag/0.5.1
 [0.5.0]: https://github.com/Browsers-software/browsers/releases/tag/0.5.0